03. advanced course/10. agent engine - WebUI/requirements.txt
google-cloud-aiplatform[langchain,reasoningengine]
google-cloud-storage
cloudpickle
streamlit